\name{findmid}
\Rdversion{0.1}
\alias{findmid}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Find the mid point of memberships of each cluster
}
\description{
Find the mid point of memberships of each cluster. Sub function of \code{\link{plot.hmac}}.
}
\usage{
findmid(x,memb)
}

\arguments{
  \item{x}{
Input data
}

  	\item{memb}{
Membership of each observation
  }
}
%
\seealso{
\code{\link{plot}} for plotting the dendrogram 
}

\author{
Surajit Ray and Yansong Cheng
}
